[skip] Extracted `find_commit` from `native_commit` {{err:8259}} (20m)
This allows us to normalize commit references and to ensure that they exist in the Repo without constructing a Houston Commit for each one.

And _that_ allows us to remove `Commit#original` which was hanging on to a reference to the Rugged commit longer than it needed.

I _thought_ that the latter was the reason that Houston was hanging on to file descriptors for too long; but that doesn't seem to be the case.