[fix] Resolved {{err:9020}} (5m)
The before_ and after_ save callbacks are executed within a transaction with Active Record's save operation.

We don't want to synchronize commits inside that transaction because a race condition with commit-syncing can result in a Unique Constraint violation. The Commit Synchronizer can recover from that, but Postgres will consider the transaction failed.