[fix] Fixed creating a project with updated version of cancancan (20m)
Not sure how this was working before, but cancancan automatically tries to create and authorize an instance of Project from the params before the create action; however, unless it knows to call the `project_attributes` method, it will try to create an instance with un-permitted parameter attributes, causing an error.

Cancancan knows to look for {resource}_params, though, so a simple alias fixes the issue.